requirements.txt
pandas
scikit-surprise
streamlit
numpy==1.26.4